requirements.txt
attrs==19.3.0
bcrypt==3.2.0
beeprint==2.4.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2
fabric==2.5.0
Flask==1.1.2
graphql-core==3.1.0
idna==2.10
importlib-metadata==1.7.0
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpath==0.82
MarkupSafe==1.1.1
more-itertools==8.4.0
numpy==1.18.5
packaging==20.4
paho-mqtt==1.5.0
paramiko==2.7.2
pendulum==2.1.2
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
sgqlc==10.1
six==1.14.0
urllib3==1.25.9
urwid==2.1.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
